+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 17.5.2
 
+### Bug Fix
+- Fixed an issue sorting forum topics by number of replies
+
 ## 17.5.1 (May 18, 2018)
 
 ### Bug Fix